pyproject.toml
[project]
name = "copilot-export"
version = "0.1.0"
description = "Export Copilot CLI sessions to Markdown and HTML"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "markdown>=3.5",
    "pygments>=2.17",
]

[project.scripts]
copilot-export = "copilot_export.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
